Refactor theme management by centralizing static themes in a new themesList.js file; update web and desktop extensions to use shared theme data
